Create reftools.html.ep
This page will display controls to request describe/text of various tables that I may need while playing. In short, a referee's tool page of random generation. It's designed so that it will be redisplayed after the description is generated, with the generated description at the top. It uses the Reactive/Bootstrap toolkit for mobile/tablet compatibility.